Add agenda for the mini-iot worshop to readme
Added the agenda for the mini-iot workshop for thursday the 17th of november.
This workshop will be of slightly smaller scope than the previous workshop.
It will instead have a stronger focus on using ifttt.com as an api for a wider
array of platforms on the web.
The agenda should probably be moved somewhere else afterwards.